Blog can be deleted only by the user who added the blog.

+++ part4/bloglist/controllers/blogs.js
@@ -67,8 +67,22 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  tokenExtractor(request);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  const userId = decodedToken.id;
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === userId) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: 'invalid user' });
+  }
 });
 
 module.exports = blogsRouter;
